requirements-tests.txt
botcity-framework-web>=0.4.0
pytest==7.0.0
pytest-html==3.1.1